+++ spotted-java-core/src/main/kotlin/dev/cjav/spotted/models/users/playlists/PlaylistCreateResponse.kt
@@ -36,7 +36,7 @@ private constructor(
     private val images: JsonField<List<ImageObject>>,
     private val name: JsonField<String>,
     private val owner: JsonField<Owner>,
-    private val public_: JsonField<Boolean>,
+    private val published: JsonField<Boolean>,
     private val snapshotId: JsonField<String>,
     private val tracks: JsonField<Tracks>,
     private val type: JsonField<String>,
@@ -65,7 +65,7 @@ private constructor(
         images: JsonField<List<ImageObject>> = JsonMissing.of(),
         @JsonProperty("name") @ExcludeMissing name: JsonField<String> = JsonMissing.of(),
         @JsonProperty("owner") @ExcludeMissing owner: JsonField<Owner> = JsonMissing.of(),
-        @JsonProperty("public") @ExcludeMissing public_: JsonField<Boolean> = JsonMissing.of(),
+        @JsonProperty("published") @ExcludeMissing published: JsonField<Boolean> = JsonMissing.of(),
         @JsonProperty("snapshot_id")
         @ExcludeMissing
         snapshotId: JsonField<String> = JsonMissing.of(),
@@ -82,7 +82,7 @@ private constructor(
         images,
         name,
         owner,
-        public_,
+        published,
         snapshotId,
         tracks,
         type,
@@ -174,7 +174,7 @@ private constructor(
      * @throws SpottedInvalidDataException if the JSON field has an unexpected type (e.g. if the
      *   server responded with an unexpected value).
      */
-    fun public_(): Optional<Boolean> = public_.getOptional("public")
+    fun published(): Optional<Boolean> = published.getOptional("published")
 
     /**
      * The version identifier for the current playlist. Can be supplied in other requests to target
@@ -279,11 +279,11 @@ private constructor(
     @JsonProperty("owner") @ExcludeMissing fun _owner(): JsonField<Owner> = owner
 
     /**
-     * Returns the raw JSON value of [public_].
+     * Returns the raw JSON value of [published].
      *
-     * Unlike [public_], this method doesn't throw if the JSON field has an unexpected type.
+     * Unlike [published], this method doesn't throw if the JSON field has an unexpected type.
      */
-    @JsonProperty("public") @ExcludeMissing fun _public_(): JsonField<Boolean> = public_
+    @JsonProperty("published") @ExcludeMissing fun _published(): JsonField<Boolean> = published
 
     /**
      * Returns the raw JSON value of [snapshotId].
@@ -343,7 +343,7 @@ private constructor(
         private var images: JsonField<MutableList<ImageObject>>? = null
         private var name: JsonField<String> = JsonMissing.of()
         private var owner: JsonField<Owner> = JsonMissing.of()
-        private var public_: JsonField<Boolean> = JsonMissing.of()
+        private var published: JsonField<Boolean> = JsonMissing.of()
         private var snapshotId: JsonField<String> = JsonMissing.of()
         private var tracks: JsonField<Tracks> = JsonMissing.of()
         private var type: JsonField<String> = JsonMissing.of()
@@ -361,7 +361,7 @@ private constructor(
             images = playlistCreateResponse.images.map { it.toMutableList() }
             name = playlistCreateResponse.name
             owner = playlistCreateResponse.owner
-            public_ = playlistCreateResponse.public_
+            published = playlistCreateResponse.published
             snapshotId = playlistCreateResponse.snapshotId
             tracks = playlistCreateResponse.tracks
             type = playlistCreateResponse.type
@@ -509,15 +509,16 @@ private constructor(
          * relevant. For more about public/private status, see
          * [Working with Playlists](/documentation/web-api/concepts/playlists)
          */
-        fun public_(public_: Boolean) = public_(JsonField.of(public_))
+        fun published(published: Boolean) = published(JsonField.of(published))
 
         /**
-         * Sets [Builder.public_] to an arbitrary JSON value.
+         * Sets [Builder.published] to an arbitrary JSON value.
          *
-         * You should usually call [Builder.public_] with a well-typed [Boolean] value instead. This
-         * method is primarily for setting the field to an undocumented or not yet supported value.
+         * You should usually call [Builder.published] with a well-typed [Boolean] value instead.
+         * This method is primarily for setting the field to an undocumented or not yet supported
+         * value.
          */
-        fun public_(public_: JsonField<Boolean>) = apply { this.public_ = public_ }
+        fun published(published: JsonField<Boolean>) = apply { this.published = published }
 
         /**
          * The version identifier for the current playlist. Can be supplied in other requests to
@@ -602,7 +603,7 @@ private constructor(
                 (images ?: JsonMissing.of()).map { it.toImmutable() },
                 name,
                 owner,
-                public_,
+                published,
                 snapshotId,
                 tracks,
                 type,
@@ -627,7 +628,7 @@ private constructor(
         images().ifPresent { it.forEach { it.validate() } }
         name()
         owner().ifPresent { it.validate() }
-        public_()
+        published()
         snapshotId()
         tracks().ifPresent { it.validate() }
         type()
@@ -659,7 +660,7 @@ private constructor(
             (images.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
             (if (name.asKnown().isPresent) 1 else 0) +
             (owner.asKnown().getOrNull()?.validity() ?: 0) +
-            (if (public_.asKnown().isPresent) 1 else 0) +
+            (if (published.asKnown().isPresent) 1 else 0) +
             (if (snapshotId.asKnown().isPresent) 1 else 0) +
             (tracks.asKnown().getOrNull()?.validity() ?: 0) +
             (if (type.asKnown().isPresent) 1 else 0) +
@@ -1445,7 +1446,7 @@ private constructor(
             images == other.images &&
             name == other.name &&
             owner == other.owner &&
-            public_ == other.public_ &&
+            published == other.published &&
             snapshotId == other.snapshotId &&
             tracks == other.tracks &&
             type == other.type &&
@@ -1464,7 +1465,7 @@ private constructor(
             images,
             name,
             owner,
-            public_,
+            published,
             snapshotId,
             tracks,
             type,
@@ -1476,5 +1477,5 @@ private constructor(
     override fun hashCode(): Int = hashCode
 
     override fun toString() =
-        "PlaylistCreateResponse{id=$id, collaborative=$collaborative, description=$description, externalUrls=$externalUrls, followers=$followers, href=$href, images=$images, name=$name, owner=$owner, public_=$public_, snapshotId=$snapshotId, tracks=$tracks, type=$type, uri=$uri, additionalProperties=$additionalProperties}"
+        "PlaylistCreateResponse{id=$id, collaborative=$collaborative, description=$description, externalUrls=$externalUrls, followers=$followers, href=$href, images=$images, name=$name, owner=$owner, published=$published, snapshotId=$snapshotId, tracks=$tracks, type=$type, uri=$uri, additionalProperties=$additionalProperties}"
 }